backend/requirements.txt
anyio==3.6.1
asgiref==3.5.2
asyncpg==0.25.0
autopep8==1.6.0
bcrypt==3.2.2
certifi==2022.6.15.1
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.3
cryptography==37.0.2
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.2.1
fastapi==0.78.0
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.1
python-gitlab==3.9.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.38
sqlalchemy2-stubs==0.0.2a24
sqlmodel==0.0.6
starlette==0.19.1
toml==0.10.2
typing_extensions==4.2.0
urllib3==1.26.12
uvicorn==0.17.6
zipp==3.8.0
